python-gateway/requirements.txt
influxdb
paho-mqtt
-e libs/influx
-e libs/docker_secrets
